Got the library working on the raspberry pi.
delay() calls were replaced with sleep_for() from the C++11 standard.
All I2C calls from the standard Arduino library with UART calls from
pigpio.

TODO: There was an issue with writing the restart command because the
UART write8() looks for the confirmation command from the BNO055, but
since the chip is told to restart it never sends that command.  A
specialized writerestart() should likely be written or the restart GPIO
pin should be used.